+++ packages/vscode-extension/CHANGELOG.md
@@ -1,3 +1,7 @@
+### 0.3.2
+* Avoid create config.env file if polyapi's library is not installed.
+* Fix polyapi's library credentials path on Windows.
+
 ### 0.3.1
 * Updated 'Too many requests' error message.